# Use Set hash in Distinct (2nd attempt)