Re-opening https://github.com/open-telemetry/opentelemetry-go/pull/5028 after new benchmarks.

```
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/otel/attribute
cpu: Intel(R) Xeon(R) CPU @ 2.20GHz
                                            │   main.txt    │              hash.txt               │
                                            │    sec/op     │    sec/op     vs base               │
EquivalentMapAccess/Empty-24                    32.01n ± 2%   10.12n ±  4%  -68.37% (p=0.002 n=6)
EquivalentMapAccess/1_string_attribute-24      106.25n ± 2%   10.01n ±  5%  -90.58% (p=0.002 n=6)
EquivalentMapAccess/10_string_attributes-24   826.250n ± 1%   9.982n ± 11%  -98.79% (p=0.002 n=6)
EquivalentMapAccess/1_int_attribute-24         106.65n ± 2%   10.13n ±  3%  -90.50% (p=0.002 n=6)
EquivalentMapAccess/10_int_attributes-24       833.25n ± 2%   10.04n ±  5%  -98.80% (p=0.002 n=6)
geomean                                         190.3n        10.06n        -94.72%
```

```
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/otel/sdk/metric
cpu: Intel(R) Xeon(R) CPU @ 2.20GHz
                                                         │   main.txt    │              hash.txt               │
                                                         │    sec/op     │    sec/op     vs base               │
SyncMeasure/NoView/Int64Counter/Attributes/0-24             264.8n ± 12%   316.1n ± 30%  +19.37% (p=0.002 n=6)
SyncMeasure/NoView/Int64Counter/Attributes/1-24             370.9n ± 25%   303.3n ± 18%  -18.24% (p=0.002 n=6)
SyncMeasure/NoView/Int64Counter/Attributes/10-24           2038.5n ± 15%   320.6n ± 17%  -84.27% (p=0.002 n=6)
SyncMeasure/NoView/Float64Counter/Attributes/0-24           237.3n ±  9%   279.5n ± 21%  +17.78% (p=0.015 n=6)
SyncMeasure/NoView/Float64Counter/Attributes/1-24           418.9n ± 17%   294.9n ± 18%  -29.61% (p=0.002 n=6)
SyncMeasure/NoView/Float64Counter/Attributes/10-24         1753.5n ±  3%   308.1n ± 13%  -82.43% (p=0.002 n=6)
SyncMeasure/NoView/Int64UpDownCounter/Attributes/0-24       277.9n ± 15%   280.7n ± 10%        ~ (p=0.937 n=6)
SyncMeasure/NoView/Int64UpDownCounter/Attributes/1-24       368.5n ± 25%   275.8n ±  7%  -25.13% (p=0.002 n=6)
SyncMeasure/NoView/Int64UpDownCounter/Attributes/10-24     1762.0n ± 14%   274.1n ± 10%  -84.45% (p=0.002 n=6)
SyncMeasure/NoView/Float64UpDownCounter/Attributes/0-24     262.1n ± 11%   281.4n ± 12%        ~ (p=0.180 n=6)
SyncMeasure/NoView/Float64UpDownCounter/Attributes/1-24     369.1n ± 14%   281.8n ± 12%  -23.68% (p=0.002 n=6)
SyncMeasure/NoView/Float64UpDownCounter/Attributes/10-24   1757.5n ± 37%   294.5n ±  8%  -83.24% (p=0.002 n=6)
SyncMeasure/NoView/Int64Histogram/Attributes/0-24           308.3n ± 10%   322.1n ± 12%        ~ (p=0.818 n=6)
SyncMeasure/NoView/Int64Histogram/Attributes/1-24           307.2n ± 12%   334.0n ± 18%        ~ (p=0.240 n=6)
SyncMeasure/NoView/Int64Histogram/Attributes/10-24         1137.0n ± 18%   359.0n ± 10%  -68.43% (p=0.002 n=6)
SyncMeasure/NoView/Float64Histogram/Attributes/0-24         316.4n ± 11%   313.6n ± 11%        ~ (p=0.818 n=6)
SyncMeasure/NoView/Float64Histogram/Attributes/1-24         354.0n ± 15%   349.6n ±  6%        ~ (p=0.699 n=6)
SyncMeasure/NoView/Float64Histogram/Attributes/10-24        941.1n ± 20%   366.2n ± 11%  -61.08% (p=0.002 n=6)
geomean                                                     533.3n         307.4n        -42.36%
```

I think we should strongly consider adopting this, or something similar. In particular, it has a very large performance impact when more than a few attributes are used.