@juliusv 2025-08-15
Some input from the sidelines (and yay on the performance improvements! 🥳):

> For reference, Prometheus has used as similar size hash for their implementation. There haven't been any known issues related to collisions there and it has been out for ~ 10 years.

It looks like collision handling was added to `client_golang` around a decade ago: https://github.com/prometheus/client_golang/pull/221

Besides the client library, we have seen some real collisions over the years in the Prometheus server that have all been addressed:

* Over a decade ago, a server with millions of series at SoundCloud ran into hash collisions. I think we used an FNV-1a hash back then to differentiate series within the TSDB. All this has completely changed by now, so the TSDB is resistant against this now.
* Another real hash collision happened more recently in the PromQL engine in Prometheus: https://github.com/prometheus/prometheus/issues/5724. But as far as I can tell, this was due to a bad custom way of how we hashed labels together.

We still use hashes without collision checking in some places, which are hopefully unlikely enough to ever cause a collision. See also https://github.com/prometheus/prometheus/issues/12519 for more discussion on hash collision potential in Prometheus.