@dashpole 2025-08-27
Yes, I will look into what was done in https://github.com/prometheus/client_golang/pull/221.