@dashpole 2025-08-27
The TL;DR is that handling collisions just means using a `map[uint64][]metricWithLabelValues` instead of a `map[uint64]metricWithLabelValues` to store aggregations (note the slice), and iterating over elements to find the exact one. Looks fairly straightforward. See https://github.com/prometheus/client_golang/blob/84a4734fe73aecc9ec36e1fbcb09c9a2d679f392/prometheus/vec.go#L319

I think the biggest difference for the attributes package is that we would need to ensure that `Equals()` correctly handles hash collisions (i.e. it should check more than that the set's Distinct are the same).  Alternatively, we could switch to the "Add a new Hash() function" approach since that keeps a correct `Equals()` implementation. I'm not sure if the current Distinct implementation has poor `==` performance or not...