@dashpole 2025-09-02
Looking a little bit more at hash collision probabilities, i'm not sure it is worth implementing hash collision resolution.

Unlike the prometheus server, the metric name isn't a label, so the name doesn't count towards cardinality for the purposes of hash collision.  So for us to reach 1 million cardinality in a `map` in the SDK would require 1 million series _per instrument_.  Even at 1 million series on a single instrument, the collision probability should be 2.5x10^-8 if hashes are random, which is quite a low chance.  For a "normal" 1000 cardinality on an instrument, it is ~2.7x10^-17...

The overhead of collision detection is high.  For the `SyncMeasure/NoView/Int64Counter/Attributes/10` case, it more than doubles the performance cost (104 ns -> 225 ns) of Measure to add collision detection.  This is from the Equals() method.